speedspacechart: group gridY with tickY as axisY layer
