+++ ui-speedspacechart/src/components/layers/index.ts
@@ -1,14 +1,12 @@
 export { default as AxisLayerX } from './AxisLayerX';
-export { default as AxisLayerY } from './AxisLayerY';
 export { default as CurveLayer } from './CurveLayer';
 export { default as DeclivityLayer } from './DeclivityLayer';
 export { default as ElectricalProfileLayer } from './ElectricalProfileLayer';
 export { default as FrontInteractivityLayer } from './FrontInteractivityLayer';
-export { default as MajorGridY } from './MajorGridY';
 export { default as PowerRestrictionsLayer } from './PowerRestrictionsLayer';
 export { default as ReticleLayer } from './ReticleLayer';
 export { default as SpeedLimitTagsLayer } from './SpeedLimitTagsLayer';
 export { default as StepLayer } from './StepLayer';
 export { default as TickLayerX } from './TickLayerX';
-export { default as TickLayerY } from './TickLayerY';
+export { default as AxisLayerY } from './TickLayerY';
 export { default as TickLayerYRight } from './TickLayerYRight';
